refactor: moving color theme to library
