+++ app/components/modules/ModuleGrid.tsx
@@ -2,7 +2,8 @@ import clsx from 'clsx';
 
 import Module from '~/components/modules/Module';
 import ProductCard from '~/components/product/Card';
-import type {SanityColorTheme, SanityModule} from '~/lib/sanity';
+import type {SanityModule} from '~/lib/sanity';
+import type {SanityColorTheme} from '~/lib/theme';
 import type {ProductWithNodes} from '~/types/shopify';
 
 // Sanity modules to render in full width (across all grid columns)
