+++ app/components/portableText/annotations/ProductInlineLink.tsx
@@ -12,7 +12,7 @@ import {Link} from '~/components/Link';
 import {AddToCartLink} from '~/components/product/buttons/AddToCartButton';
 import BuyNowButton from '~/components/product/buttons/BuyNowButton';
 import ProductTooltip from '~/components/product/Tooltip';
-import type {SanityColorTheme} from '~/lib/sanity';
+import type {SanityColorTheme} from '~/lib/theme';
 import type {ProductWithNodes} from '~/types/shopify';
 
 type Props = {
