+++ app/lib/theme.ts
@@ -1,0 +1,4 @@
+export type SanityColorTheme = {
+  background: string;
+  text: string;
+};
